update weapon class to use plasma gun scriptable object

+++ Weapons/PlasmaGun/Bullet.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public class Bullet : MonoBehaviour {
-    public int damage = 1;                                           // Damage caused by this bullet.
+    public float damage = 1;                                           // Damage caused by this bullet.
     public float maximunDistance = 150f;                             // Maximun distance the bullet can move towards before being restore to the bullet object pool.
     public GameObject impactParticles;                               // GameObject which contains impact particle effect.             
     private Vector3 _originalPosition;                               // Original bullet position. Used to restore bullet to weapoin shooting origin after the bullet is destroyed.
